Move comparison code to a new file
