+++ mira/dkg/wsgi.py
@@ -13,8 +13,7 @@ from mira.dkg.grounding import grounding_blueprint
 from mira.dkg.model import model_blueprint
 from mira.dkg.ui import ui_blueprint
 from mira.dkg.utils import PREFIXES, MiraState
-from mira.metamodel.templates import RefinementClosure
-
+from mira.metamodel.comparison import RefinementClosure
 
 logger = logging.getLogger(__name__)
